bug fixes found in lab testing.
